+++ src/commands/queue.ts
@@ -5,6 +5,7 @@ import {TYPES} from '../types.js';
 import PlayerManager from '../managers/player.js';
 import Command from './index.js';
 import {buildQueueEmbed} from '../utils/build-embed.js';
+import {getGuildSettings} from '../utils/get-guild-settings.js';
 
 @injectable()
 export default class implements Command {
@@ -17,7 +18,9 @@ export default class implements Command {
       .setRequired(false))
     .addIntegerOption(option => option
       .setName('page-size')
-      .setDescription('how many items to display per page [default: 10]')
+      .setDescription('how many items to display per page [default: 10, max: 50]')
+      .setMinValue(1)
+      .setMaxValue(50)
       .setRequired(false));
 
   private readonly playerManager: PlayerManager;
@@ -27,12 +30,16 @@ export default class implements Command {
   }
 
   public async execute(interaction: ChatInputCommandInteraction) {
-    const player = this.playerManager.get(interaction.guild!.id);
+    const guildId = interaction.guild!.id;
+    const player = this.playerManager.get(guildId);
+
+    const pageSizeFromOptions = interaction.options.getInteger('page-size');
+    const pageSize = pageSizeFromOptions ?? (await getGuildSettings(guildId)).defaultQueuePageSize;
 
     const embed = buildQueueEmbed(
       player,
       interaction.options.getInteger('page') ?? 1,
-      interaction.options.getInteger('page-size') ?? 10,
+      pageSize,
     );
 
     await interaction.reply({embeds: [embed]});